// src/app/data/model/dataTestAleatorio.ts
export interface DataTestAleatorio {
    cdicurso: number;
    id_curso: string;
    nombre_curso: string;
    cdipermiso: number;
    nombre_permiso: string;
    cdicategoria: number;
    nombre_categoria: string;
    descripcion_categoria: string;
    traducir: number;
    idioma: string;
    ayuda: number;
    autocorreccion: number;
    tipo_test: number;
}